Restructured component directories and created component-level CSS
