+++ src/components/Card/Card.js
@@ -1,6 +1,8 @@
+import styles from "./Card.module.css";
+
 export default function Card({ emojis, key }) {
   return (
-    <div className="Card" key={key}>
+    <div className={styles.Card} key={key}>
       {emojis.map((emoji) => emoji)}
     </div>
   );
