+++ src/components/Card/Card.module.css
@@ -1,0 +1,11 @@
+.Card {
+    width: 100px;
+    height: 100px;
+    border: 1px solid black;
+    margin: 5px;
+    background-color: white;
+    border-radius: 100px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
