+++ src/components/Deck/Deck.module.css
@@ -1,0 +1,7 @@
+.Deck {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: left;
+    margin: 10px;
+}
