--- src/components/Header.js
@@ -1,10 +1,0 @@
-export default function Header() {
-  return (
-    <header id="Header">
-      <div id="header-content">
-        <img src="./logo_100px.png" alt="Monomoji Logo" id="logo" />
-        <h1>MONOMOJI</h1>
-      </div>
-    </header>
-  );
-}
