--- src/components/Randomize.js
@@ -1,7 +1,0 @@
-export default function Randomize({ randomizeEmojis }) {
-  return (
-    <button id="Randomize" type="button" onClick={randomizeEmojis}>
-      RANDOMIZE 🔀
-    </button>
-  );
-}
