+++ src/components/Randomize/Randomize.module.css
@@ -1,0 +1,9 @@
+.Randomize {
+    border-radius: 20px;
+    text-align: center;
+    width: unset;
+}
+
+.Test {
+    background-color: blue;
+}
