+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import Header from "../components/Header";
 import Info from "../components/Info";
-import DeckGenerator from "../components/DeckGenerator";
+import DeckGenerator from "../components/DeckGeneratorr";
 
 import Head from "next/head";
