+++ src/styles/global.css
@@ -11,37 +11,11 @@ body {
   background: white;
 }
 
-header {
-  width: 100%;
-  height: 100px;
-  background-color: black;
-}
-
-#header-content {
-  width: 90%;
-  max-width: 900px;
-  margin: 0 auto;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  flex-direction: row;
-}
-
-#logo {
-  height: 100px;
-  aspect-ratio: 1 / 1;
-  padding: 10px
-}
-
-#header-content > h1 {
+h1 {
   text-align: center;
   font-weight: 700;
   font-size: 3rem;
   color: white;
-}
-
-a {
-  padding: 0;
 }
 
 main {
@@ -85,7 +59,7 @@ select {
   outline: none;
 }
 
-label > p {
+label>p {
   display: inline;
 }
 
@@ -120,30 +94,4 @@ button {
   margin-bottom: 0;
   width: 20%;
   min-width: 10rem;
-}
-
-.Card {
-  width: 100px;
-  height: 100px;
-  border: 1px solid black;
-  margin: 5px;
-  background-color: white;
-  border-radius: 100px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-#Deck {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: left;
-  margin: 10px;
-}
-
-#Randomize {
-  border-radius: 20px;
-  text-align: center;
-  width: unset;
 }
